nonebot_plugin_tieba_monitor/requirements.txt
aiotieba>=1.4.0
openai>=1.0.0
nonebot2>=2.1.0
nonebot-adapter-onebot>=2.2.3
nonebot-plugin-apscheduler>=0.3.0
